updated package and group name
