+++ build.gradle
@@ -11,7 +11,7 @@ buildscript {
 }
 
 version "0.1"
-group "com.lakerpolling"
+group "edu.oswego.cs.lakerpolling"
 
 apply plugin: "eclipse"
 apply plugin: "idea"
