+++ grails-app/controllers/edu/oswego/cs/lakerpolling/UrlMappings.groovy
@@ -1,0 +1,17 @@
+package edu.oswego.cs.lakerpolling
+
+class UrlMappings {
+
+    static mappings = {
+        "/$controller/$action?/$id?(.$format)?" {
+            constraints {
+                // apply constraints here
+            }
+        }
+
+    "/"(controller: 'dashboard', action: 'index')
+
+        "500"(view: '/error')
+        "404"(view: '/notFound')
+    }
+}
