--- grails-app/controllers/edu/oswego/edu/cs/lakerpolling/UrlMappings.groovy
@@ -1,17 +1,0 @@
-package edu.oswego.edu.cs.lakerpolling
-
-class UrlMappings {
-
-    static mappings = {
-        "/$controller/$action?/$id?(.$format)?" {
-            constraints {
-                // apply constraints here
-            }
-        }
-
-    "/"(controller: 'dashboard', action: 'index')
-
-        "500"(view: '/error')
-        "404"(view: '/notFound')
-    }
-}
