--- grails-app/controllers/edu/oswego/edu/cs/lakerpolling/controllers/LogoutController.groovy
@@ -1,19 +1,0 @@
-package edu.oswego.edu.cs.lakerpolling.controllers
-
-import grails.converters.JSON
-
-class LogoutController {
-
-    def authenticatorService
-
-    /**
-     * Endpoint to invalidate the server's tracking of this session.
-     */
-    def invalidateSignin() {
-        authenticatorService.signOutCurrentUser(session)
-        render ([
-                "message" : "ok"
-        ] as JSON)
-    }
-
-}
