+++ grails-app/domain/edu/oswego/cs/lakerpolling/models/authentication/OAuthId.groovy
@@ -1,0 +1,16 @@
+package edu.oswego.cs.lakerpolling.models.authentication
+
+class OAuthId  {
+
+    String subject
+
+    static belongsTo = [user : User]
+
+    static mapping = {
+        subject unique: true
+        autoTimestamp false
+        version false
+    }
+
+
+}
