+++ grails-app/domain/edu/oswego/cs/lakerpolling/models/authentication/User.groovy
@@ -1,0 +1,21 @@
+package edu.oswego.cs.lakerpolling.models.authentication
+
+class User {
+
+    String firstName
+    String lastName
+    String email
+    String imageUrl
+    String subject
+
+    static mapping = {
+        version false
+    }
+
+    static constraints = {
+        email email: true, unique: true
+        firstName blank: false
+        lastName blank:false
+    }
+
+}
