--- grails-app/domain/edu/oswego/edu/cs/lakerpolling/models/authentication/OAuthId.groovy
@@ -1,16 +1,0 @@
-package edu.oswego.edu.cs.lakerpolling.models.authentication
-
-class OAuthId  {
-
-    String subject
-
-    static belongsTo = [user : User]
-
-    static mapping = {
-        subject unique: true
-        autoTimestamp false
-        version false
-    }
-
-
-}
