--- grails-app/domain/edu/oswego/edu/cs/lakerpolling/models/authentication/User.groovy
@@ -1,21 +1,0 @@
-package edu.oswego.edu.cs.lakerpolling.models.authentication
-
-class User {
-
-    String firstName
-    String lastName
-    String email
-    String imageUrl
-    String subject
-
-    static mapping = {
-        version false
-    }
-
-    static constraints = {
-        email email: true, unique: true
-        firstName blank: false
-        lastName blank:false
-    }
-
-}
