+++ grails-app/init/edu/oswego/cs/lakerpolling/BootStrap.groovy
@@ -1,0 +1,14 @@
+package edu.oswego.cs.lakerpolling
+
+class BootStrap {
+
+    def init = { servletContext ->
+
+    }
+
+
+    def destroy = {
+
+    }
+
+}
