--- grails-app/init/edu/oswego/edu/cs/lakerpolling/BootStrap.groovy
@@ -1,14 +1,0 @@
-package edu.oswego.edu.cs.lakerpolling
-
-class BootStrap {
-
-    def init = { servletContext ->
-
-    }
-
-
-    def destroy = {
-
-    }
-
-}
