+++ src/main/groovy/edu/oswego/cs/lakerpolling/util/Pair.groovy
@@ -1,0 +1,14 @@
+package edu.oswego.cs.lakerpolling.util
+
+
+class Pair<K, V> {
+
+    K key
+    V value
+
+    Pair(K key, V value) {
+        this.key = key
+        this.value = value
+    }
+
+}
