--- src/main/groovy/edu/oswego/edu/cs/lakerpolling/util/Pair.groovy
@@ -1,14 +1,0 @@
-package edu.oswego.edu.cs.lakerpolling.util
-
-
-class Pair<K, V> {
-
-    K key
-    V value
-
-    Pair(K key, V value) {
-        this.key = key
-        this.value = value
-    }
-
-}
