+++ src/test/groovy/edu/oswego/cs/lakerpolling/controllers/LoginControllerSpec.groovy
@@ -1,0 +1,22 @@
+package edu.oswego.cs.lakerpolling.controllers
+
+import grails.test.mixin.TestFor
+import spock.lang.Specification
+
+/**
+ * See the API for {@link grails.test.mixin.web.ControllerUnitTestMixin} for usage instructions
+ */
+@TestFor(LoginController)
+class LoginControllerSpec extends Specification {
+
+    def setup() {
+    }
+
+    def cleanup() {
+    }
+
+    void "test something"() {
+        expect:"fix me"
+            true == false
+    }
+}
